fix: decouple sqlite from gherkin and executor
