+++ packages/gherkin/package.json
@@ -46,7 +46,6 @@
     "@cucumber/cucumber-expressions": "^18.1.0",
     "@cucumber/gherkin": "^37.0.1",
     "@cucumber/messages": "^31.2.0",
-    "@letsrunit/store": "workspace:*",
     "@letsrunit/utils": "workspace:*"
   },
   "devDependencies": {
