+++ packages/gherkin/src/index.ts
@@ -1,6 +1,7 @@
 export * from './parameters';
 export type { KeyCombo } from './keys/parse-key-combo';
 export * from './feature';
+export * from './ids';
 export * from './sanitize';
 export * from './normalize-step';
 export * from './normalize-steps';
